TOFOLAT

+++ tapuilibrary/src/main/java/company/tap/tapuilibrary/atoms/TapTextInput.kt
@@ -265,8 +265,8 @@ open class TapTextInput  @JvmOverloads constructor(
 
     override fun setTheme(theme: TextInputTheme) {
         theme.textColor?.let { setTextColor(it) }
-        theme.textSize?.let { textSize = it }
-        theme.letterSpacing?.let { letterSpacing = it }
+        theme.textSize?.let { textSize = it.toFloat() }
+        theme.letterSpacing?.let { letterSpacing = it.toFloat() }
         invalidate()
     }
 }
