+++ tapuilibrary/src/main/java/company/tap/tapuilibrary/atoms/TapTextView.kt
@@ -31,8 +31,8 @@ open class TapTextView(context: Context, attributeSet: AttributeSet?) :
      **/
     override fun setTheme(theme: TextViewTheme) {
         theme.textColor?.let { setTextColor(it) }
-        theme.textSize?.let { textSize = it }
-        theme.letterSpacing?.let { letterSpacing = it }
+        theme.textSize?.let { textSize = it.toFloat() }
+        theme.letterSpacing?.let { letterSpacing = it.toFloat() }
         invalidate()
     }
 }
